optimize the test structure
