+++ LambdaAndJdkCollections/src/test/java/org/sample/lambda/PopulateTest.java
@@ -1,0 +1,17 @@
+package org.sample.lambda;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+import org.junit.Before;
+
+public class PopulateTest {
+
+	protected List<String> sampleStringList;
+
+	@Before
+	public void beforeTest() {
+		sampleStringList = new ArrayList<>(Arrays.asList("You", "must", "be", "a", "professional", "guy"));
+	}
+}
