+++ LambdaAndJdkCollections/src/test/java/org/sample/lambda/ReplaceElementTest.java
@@ -10,14 +10,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-public class ReplaceElementTest {
-
-	private List<String> sampleStringList;
-
-	@Before
-	public void beforeTest() {
-		sampleStringList = new ArrayList<>(Arrays.asList("You", "must", "be", "a", "professional", "guy"));
-	}
+public class ReplaceElementTest extends PopulateTest{
 
 	@Test
 	public void replaceElementForApacheCommons() {
